chore(server): refactor naming of application module
