+++ server/src/storage/storage.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common'
 import { BucketsController } from '../storage/buckets.controller'
 import { CoreModule } from '../core/core.module'
-import { ApplicationsModule } from '../application/applications.module'
+import { ApplicationModule } from '../application/application.module'
 import { StorageService } from './storage.service'
 
 @Module({
-  imports: [CoreModule, ApplicationsModule],
+  imports: [CoreModule, ApplicationModule],
   controllers: [BucketsController],
   providers: [StorageService],
 })
